+++ app/src/main/java/domain/models/Result.kt
@@ -3,6 +3,6 @@ package domain.models
 data class Result(
     var testName: String,
     var userName: String,
-    var points: Double,
+    var points: Int,
     var date: String
 )
